Doing refactoring on duplicated code from Goaba and Dan
We got rid of nested "if"s and extracted 2 methods out of
updContactStatus() and updContactState(). We were considering to also
use the extracted method called updateContent() in updateContactName(),
but we couldn't do it, due to that ecTmp.setName(strName) would also
pass a string just as updateContactStatus() does. We also thought about
to use a strategy pattern to deal the three very similar methods
(updContactName(). updContactState(). and updContactStatus()), but we
gave up on the thought as doing this would still leave a lot of
duplicated code, which doesn't solve our problem here.